ihklib: add setup functions using string list for krm
Following functions are added:
ihk_reserve_cpu_str
ihk_reserve_mem_conf_str
ihk_reserve_mem_str
ihk_os_set_ikc_map_str
ihk_os_kargs_str

+++ test/ihklib/blackbox/include/okng.h
@@ -28,6 +28,6 @@
 		goto out;					\
 	} \
 } while (0)
-#define ARRAY_SIZE_CHECK(array, size) INTERR(sizeof(array)/sizeof(array[0]) != size, "size of array \"%s\" isn't %d\n", #array, size)
+#define ARRAY_SIZE_CHECK(array, size) INTERR(sizeof(array)/sizeof(array[0]) != size, "size of array \"%s\" isn't %d\n", #array, (int)size)
 
 #endif
